Rename project to architect
