+++ architect/components/lenses.py
@@ -8,8 +8,8 @@ import astropy.constants as const
 import numpy as np
 
 # project
-from payload_designer.components import Component
-from payload_designer.luts import LUT
+from architect.components import Component
+from architect.luts import LUT
 
 LOG = logging.getLogger(__name__)
