+++ architect/luts/sensors/tauswir_quantum_efficiency.csv
