+++ architect/pipelines_LEGACY/grism_eff_LEGACY.py
@@ -11,8 +11,8 @@ import numpy as np
 import pandas as pd
 
 # project
-from payload_designer.components import diffractors
-from payload_designer.libs import plotlib
+from architect.components import diffractors
+from architect.libs import plotlib
 
 # region path config
 filename = Path(__file__).stem
