+++ architect/systems/system.py
@@ -5,8 +5,8 @@ import numpy as np
 import pandas as pd
 
 # project
-from payload_designer import Artifact
-from payload_designer.luts import LUT
+from architect import Artifact
+from architect.luts import LUT
 
 
 class System(Artifact):
