+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "payload_designer"
+name = "architect"
 version = "0.1.0"
 description = ""
 authors = ["UTAT"]
@@ -47,7 +47,7 @@ wrap-summaries = 80
 [tool.pylint.master]
 ignore-patterns="^((?!\\.py).)*$"  # ignore files that are not python files
 recursive="yes"
-disable = ["invalid-name","too-many-arguments", "too-many-locals", "arguments-differ", "unnecessary-lambda", "too-many-instance-attributes", "protected-access", "duplicate-code", "too-few-public-methods", "logging-fstring-interpolation", "not-callable", "too-many-statements", "pointless-statement", "not-context-manager","fixme","too-many-branches","too-many-lines","line-too-long","anomalous-backslash-in-string","too-many-function-args"]
+disable = ["invalid-name","too-many-arguments", "too-many-locals", "arguments-differ", "unnecessary-lambda", "too-many-instance-attributes", "protected-access", "duplicate-code", "too-few-public-methods", "logging-fstring-interpolation", "not-callable", "too-many-statements", "pointless-statement", "not-context-manager","fixme","too-many-branches","too-many-lines","line-too-long","anomalous-backslash-in-string","too-many-function-args","missing-final-newline"]
 max-line-length = 88
 generated-members=["numpy.*","torch.*","pygame.*"] # https://github.com/pytorch/pytorch/issues/701
