+++ tests/test_components/test_diffractors/test_VPHGrating.py
@@ -6,7 +6,7 @@ import logging
 import pytest
 
 # project
-from payload_designer.components import diffractors
+from architect.components import diffractors
 
 LOG = logging.getLogger(__name__)
