+++ tests/test_components/test_foreoptics/test_Foreoptic.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 
 # project
-from payload_designer.components import foreoptics
+from architect.components import foreoptics
 
 LOG = logging.getLogger(__name__)
