+++ tests/test_components/test_lenses/test_ThinLens.py
@@ -8,7 +8,7 @@ import numpy as np
 import pytest
 
 # project
-from payload_designer.components import lenses
+from architect.components import lenses
 
 LOG = logging.getLogger(__name__)
